Factor out weights into separate files
