+++ includes/weights/phase_count.hpp
@@ -1,0 +1,8 @@
+# pragma once
+
+const int pc_pawn = 2;
+const int pc_knight = 5;
+const int pc_bishop = 13;
+const int pc_rook = 17;
+const int pc_queen = 40;
+const int pc_intercept = -1;
